requirements.txt
pillow
django-widget-tweaks
ipython